Revert "Member registration"
