+++ app/register/page.tsx
@@ -1,5 +1,7 @@
 import Link from "next/link";
 
+import RegisterComponent from "./member/page";
+
 export default function Register() {
   return (
     <div className="container flex flex-col justify-center p-60">
